requirement.txt
asgiref==3.8.1
certifi==2024.2.2
cloudinary==1.39.1
Django==4.2.11
djangorestframework==3.15.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
